requirements.txt
torch
torch-cosine-annealing
tqdm
pyyaml
tensorboard
optuna
numpy